Refactor playback logic and unify server URL handling.
Updated playback methods to use a unified `serverUrl` parameter, improving code clarity and reducing duplication. Replaced the `long` types with `decimal` for position/length calculations for higher precision. Removed unused or deprecated pause/resume/stop methods across multiple files to simplify audio playback management.

+++ pMusic/Audio/IAudioBackend.cs
@@ -4,8 +4,8 @@ namespace pMusic.Interface;
 
 public interface IAudioBackend
 {
-    long GetPosition(int stream);
-    long GetLength(int stream);
+    decimal GetPosition(int stream);
+    decimal GetLength(int stream);
 
     PlaybackState GetState(int stream);
 }
